services/diagnostico-service/requirements.txt
fastapi
uvicorn
sqlalchemy
numpy
psycopg2-binary
python-dotenv
httpx
Pillow
tensorflow-cpu
python-multipart